movable camera

+++ src/Utils/Math.h
@@ -58,7 +58,7 @@ inline float reflectance(float cosine, float refractionRatio) {
     // schlick's reflectance approximation
     auto r0 = (1 - refractionRatio) / (1 + refractionRatio);
     r0 *= r0;
-    return r0 + (1 - r0) * std::pow((1 - cosine), 5);
+    return r0 + (1 - r0) * static_cast<float>(std::pow((1 - cosine), 5));
 }
 
 template <typename T>
